llm-service/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3

# LangChain and LangGraph
langchain==0.1.0
langgraph==0.0.26
langchain-openai==0.0.5
langchain-anthropic==0.0.1

# Additional utilities
httpx==0.26.0
python-multipart==0.0.6

# Development dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
